Add tasks for modularity and debugging
Instead of creating a title method, I plan to implement the debug
message system in DBReader into BFFrameAnimator so that users can use
the (to be implemented) debug mode to pinpoint issues.

The modularity task is mostly for units who share files with other
similar units (e.g. metal ghosts).